Sort additions -In hppc.sorting, added template-generated sort classes for primitives and generics alike. -Added template-generated related Unit tests, by porting IndirectSort certification tests -For now, implemented in-place dual-pivot QuickSort by [Yaroslavskiy 2009] like in IndirectSort -sort methods take both Comparable and Comparator sort functions to look like java.util.Arrays ones

+++ hppc-core/src/test/templates/com/carrotsearch/hppc/KTypeVTypeOpenHashMapTest.java
@@ -84,10 +84,10 @@ public class KTypeVTypeOpenHashMapTest<KType, VType> extends AbstractKTypeTest<K
      * Convert a VType to long, (VType being a boxed elementary type or a primitive), else
      * returns 0L.
      */
-    protected long vcastType(VType type)
+    protected int vcastType(VType type)
     {
         /*! #if ($TemplateOptions.VTypePrimitive)
-        return (long) type;
+        return (int) type;
         #else !*/
         long k = 0L;
 
@@ -100,7 +100,7 @@ public class KTypeVTypeOpenHashMapTest<KType, VType> extends AbstractKTypeTest<K
             k = ((Number) type).longValue();
         }
 
-        return k;
+        return (int) k;
         /*! #end !*/
     }
